% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Festem-package.R
\docType{package}
\name{EMFS}
\alias{EMFS-package}
\alias{EMFS}
\title{EMFS: Feature screening for clustering analysis}
\description{
Feature screening for clustering analysis of count data with an application to single-cell RNA-sequencing. See Wang, Chen and Xi (2023) <https://doi.org/10.48550/arXiv.2306.12671> for more details.
}
\author{
\strong{Maintainer}: Zihao Chen \email{g.e.challenger@pku.edu.cn} (\href{https://orcid.org/0000-0002-1382-0409}{ORCID})

Authors:
\itemize{
  \item Changhu Wang \email{wangch156@pku.edu.cn} (\href{https://orcid.org/0009-0002-8567-0961}{ORCID})
  \item Ruibin Xi \email{ruibinxi@math.pku.edu.cn} (\href{https://orcid.org/0000-0001-7545-7361}{ORCID}) [copyright holder]
}

}
